backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
google-cloud-datastore==2.20.0
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
pytz==2024.1
